+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/util/logs/timber/NoLogTree.kt
@@ -1,4 +1,4 @@
-package net.nymtech.nymvpn.util.timber
+package net.nymtech.nymvpn.util.logs.timber
 
 import timber.log.Timber